Updated SCC test to be invoked twice
The problem with the sharedClasses test right now is
that even though there are 5 JVMs invoked, which means
that some JVMs will load AOT code compiled by a different
JVM, it doesn't fully test the AOT load infrastructure
becuase after the test is completed, the SCC gets
destroyed before the next test runs. This commit addresses
this issue by invoking the test twice; once to primarily
populate the SCC, and a second time to primarily load
from it.

https://github.com/eclipse/openj9/issues/3774

Signed-off-by: Irwin D'Souza <dsouzai@ca.ibm.com>